requirements.txt
tldextract
colorama
requests
chardet
ua_generator
pycryptodome